fix: add thread button story book and the config
