+++ .storybook/main.js
@@ -12,24 +12,5 @@ const config = {
     name: "@storybook/react-vite",
     options: {},
   },
-  webpackFinal: async (config) => {
-    config.module.rules.push({
-      test: /\,css&/,
-      use: [
-        {
-          loader: "postcss-loader",
-          options: {
-            ident: "postcss",
-            plugins: [require("tailwindcss"), require("autoprefixer")],
-          },
-        },
-      ],
-      include: path.resolve(__dirname, "../"),
-    });
-    return config;
-  },
-  core: {
-    builder: "webpack5",
-  },
 };
 export default config;
