+++ src/components/AddThreadButton.jsx
@@ -1,11 +1,11 @@
 import PropTypes from "prop-types";
 
 const AddThreadButton = (props) => {
-  const { label, handleShowModal } = props;
+  const { label, handleShowModal, className } = props;
   return (
     <div>
       <button
-        className="btn h-fit min-h-fit w-full py-1 text-lg rounded-md bg-blue-400 text-white hover:bg-blue-500"
+        className={`btn h-fit min-h-fit w-full py-1 text-lg rounded-md text-white ${className ? className : "bg-blue-400 hover:bg-blue-500"}`}
         onClick={handleShowModal}
       >
         {label}
@@ -16,6 +16,7 @@ const AddThreadButton = (props) => {
 
 AddThreadButton.propTypes = {
   label: PropTypes.string.isRequired,
+  className: PropTypes.string.isRequired,
   handleShowModal: PropTypes.func.isRequired,
 };
